Fix ghost income (#1061)
* fix reconnect resurrecting player

* use uint16 for money

* dereference ghost player

+++ app/models/colyseus-models/player.ts
@@ -31,7 +31,7 @@ export default class Player extends Schema implements IPlayer {
   @type(["string"]) shop = new ArraySchema<Pkm>()
   @type(ExperienceManager) experienceManager = new ExperienceManager()
   @type({ map: "uint8" }) synergies = new Synergies()
-  @type("uint8") money = process.env.MODE == "dev" ? 400 : 6
+  @type("uint16") money = process.env.MODE == "dev" ? 400 : 6
   @type("uint8") life = 100
   @type("boolean") shopLocked: boolean = false
   @type("uint8") streak: number = 0
